# idt/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: IDT API
  version: "1.0.0"
  description: |
    FastAPI + LangGraph/LangChain 기반 RAG & Agent 시스템 API

    **아키텍처**: Thin DDD (Domain → Application → Infrastructure)

    **주요 기능**:
    - PDF/Excel 문서 업로드 및 벡터 색인
    - BM25 + Vector 하이브리드 검색
    - ReAct RAG Agent 기반 질의응답
    - Multi-Turn 대화 메모리 관리
    - Agent Builder (Custom / Middleware / Auto)
    - MCP 서버 동적 등록 관리
servers:
  - url: http://localhost:8000
    description: Local Development
...
paths:
...
  /api/v1/documents/upload:
    post:
      tags: [documents]
      summary: PDF 문서 업로드
      description: |
        PDF를 LangGraph 워크플로우로 처리합니다.
        파싱 → 카테고리 분류 → Parent/Child 청킹 → Qdrant 벡터 저장
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
          description: 문서 소유자 ID
        - name: child_chunk_size
          in: query
          required: false
          schema:
            type: integer
            default: 500
            minimum: 100
            maximum: 4000
          description: Child 청크 크기 (토큰, 100–4000)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: 업로드할 PDF 파일
      responses:
        "200":
          description: 처리 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentUploadResponse"
              example:
                document_id: "doc-uuid-1234"
                filename: "policy.pdf"
                category: "finance"
                category_confidence: 0.92
                total_pages: 10
                chunk_count: 42
                stored_ids: ["id1", "id2"]
                status: "success"
                errors: []
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/documents/upload/async:
    post:
      tags: [documents]
      summary: PDF 비동기 업로드 (큐 등록)
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: 큐 등록 성공
          content:
            application/json:
              example:
                task_id: "task-uuid-5678"
                status: "pending"
                message: "Document queued for processing"
  /api/v1/documents/upload/status/{task_id}:
    get:
      tags: [documents]
      summary: 비동기 업로드 상태 조회
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 작업 상태
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsyncTaskStatus"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Ingest ─────────────────────────────────────────────────────────────
...
  /api/v1/doc-chunk/upload:
    post:
      tags: [doc-chunk]
      summary: 파일 청킹 결과 미리보기
      description: |
        PDF/Excel/TXT/MD 파일을 업로드하고 청킹 결과만 반환합니다.
        벡터 저장 없이 청크를 확인할 때 사용합니다.
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
        - name: strategy_type
          in: query
          schema:
            type: string
            default: parent_child
            enum: [full_token, parent_child, semantic]
        - name: chunk_size
          in: query
          schema:
            type: integer
            default: 500
            minimum: 100
            maximum: 8000
        - name: chunk_overlap
          in: query
          schema:
            type: integer
            default: 50
            minimum: 0
            maximum: 500
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: ".pdf/.xlsx/.xls/.txt/.md"
      responses:
        "200":
          description: 청킹 결과
        "422":
          $ref: "#/components/responses/ValidationError"

  # ── Excel Upload ───────────────────────────────────────────────────────
  /api/v1/excel/upload:
    post:
      tags: [excel]
      summary: Excel 파일 업로드 및 벡터 저장
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
        - name: strategy_type
          in: query
          schema:
            type: string
            default: full_token
            enum: [full_token, parent_child, semantic]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: ".xlsx 또는 .xls 파일"
      responses:
        "200":
          description: 처리 성공
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Excel Export ───────────────────────────────────────────────────────
...
  /api/v1/chunk-index/upload:
    post:
      tags: [chunk-index]
      summary: 텍스트 청킹 + 키워드 추출 + ES 색인
      description: |
        텍스트를 청킹하고 빈도 기반 키워드를 추출하여
        Elasticsearch에 색인합니다. BM25 하이브리드 검색 품질 향상용.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChunkIndexAPIRequest"
            example:
              document_id: "doc-001"
              content: "금융투자업자는 투자자의 이익을 최우선으로..."
              user_id: "user-001"
              strategy_type: "parent_child"
              chunk_size: 500
              chunk_overlap: 50
              top_keywords: 10
      responses:
        "200":
          description: 색인된 청크 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChunkIndexAPIResponse"
        "422":
          $ref: "#/components/responses/ValidationError"

  # ── Morph Index ────────────────────────────────────────────────────────
  /api/v1/morph-index/upload:
    post:
      tags: [morph-index]
      summary: Kiwi 형태소 분석 + Qdrant + ES 이중 색인
      description: |
        Kiwi 형태소 분석(NNG/NNP/VV원형/VA원형)으로 키워드를 추출하고
        Qdrant(벡터)와 Elasticsearch(BM25)에 동시 색인합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MorphIndexAPIRequest"
            example:
              document_id: "doc-001"
              content: "한국 금융시장에서 투자자 보호를 위한 규정..."
              user_id: "user-001"
              strategy_type: "parent_child"
              chunk_size: 500
              chunk_overlap: 50
              source: "financial_policy.pdf"
      responses:
        "200":
          description: 이중 색인 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MorphIndexAPIResponse"
        "422":
          $ref: "#/components/responses/ValidationError"

  # ── RAG Agent ──────────────────────────────────────────────────────────
...
  /api/v1/agents:
    post:
      tags: [Agent Builder]
      summary: 에이전트 생성
      description: LLM이 도구 자동 선택 + 시스템 프롬프트 자동 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
      responses:
        "201":
          description: 에이전트 생성 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAgentResponse"
  /api/v1/agents/{agent_id}:
    get:
      tags: [Agent Builder]
      summary: 에이전트 조회
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 에이전트 정의
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Agent Builder]
      summary: 에이전트 시스템 프롬프트/이름 수정
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentRequest"
      responses:
        "200":
          description: 수정 성공
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /api/v2/agents:
    post:
      tags: [middleware-agent]
      summary: 미들웨어 에이전트 생성
      description: |
        LangChain Middleware 기반 에이전트를 생성합니다.
        지원 미들웨어: SummarizationMiddleware, PIIMiddleware,
        ToolRetryMiddleware, ModelCallLimitMiddleware, ModelFallbackMiddleware
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMiddlewareAgentRequest"
      responses:
        "201":
          description: 에이전트 생성 성공
  /api/v2/agents/{agent_id}:
    get:
      tags: [middleware-agent]
      summary: 미들웨어 에이전트 조회
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
        - name: request_id
          in: query
          schema:
            type: string
            default: no-request-id
      responses:
        "200":
          description: 에이전트 정의
    patch:
      tags: [middleware-agent]
      summary: 미들웨어 에이전트 수정
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMiddlewareAgentRequest"
      responses:
        "200":
          description: 수정 성공
...
  /api/v3/agents/auto:
    post:
      tags: [auto-agent-builder]
      summary: 자연어 기반 자동 에이전트 빌드 시작
      description: |
        자연어 요청으로 에이전트를 자동 생성합니다.
        LLM이 도구+미들웨어를 자동 추론하고,
        보충 질문이 필요하면 세션 ID를 반환합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoBuildRequest"
            example:
              user_request: "금융 문서를 검색하고 할루시네이션을 검증하는 에이전트를 만들어줘"
              user_id: "user-001"
      responses:
        "202":
          description: 에이전트 생성 또는 보충 질문 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoBuildResponse"
  /api/v3/agents/auto/{session_id}/reply:
    post:
      tags: [auto-agent-builder]
      summary: 보충 질문 답변 제출 → 재추론 → 에이전트 생성
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoBuildReplyRequest"
      responses:
        "200":
          description: 에이전트 생성 결과
  /api/v3/agents/auto/{session_id}:
    get:
      tags: [auto-agent-builder]
      summary: 빌드 세션 상태 조회
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 세션 상태
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoBuildSessionStatusResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── MCP Registry ───────────────────────────────────────────────────────
...
components:
  responses:
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          example:
            detail: "Not found"
    ValidationError:
      description: 유효성 검사 실패
      content:
        application/json:
          example:
            detail: "strategy_type must be one of: ['full_token', 'parent_child', 'semantic']"
    InternalError:
      description: 서버 내부 오류
      content:
        application/json:
          example:
            detail: "Internal server error"
  schemas:
    DocumentUploadResponse:
      type: object
      properties:
        document_id:
          type: string
        filename:
          type: string
        category:
          type: string
          nullable: true
        category_confidence:
          type: number
          format: float
        total_pages:
          type: integer
        chunk_count:
          type: integer
        stored_ids:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [success, failed]
        errors:
          type: array
          items:
            type: string
    AsyncTaskStatus:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [pending, processing, success, failed]
        result:
          $ref: "#/components/schemas/DocumentUploadResponse"
          nullable: true
        error:
          type: string
          nullable: true
...
    ChunkIndexAPIRequest:
      type: object
      required: [document_id, content, user_id]
      properties:
        document_id:
          type: string
        content:
          type: string
        user_id:
          type: string
        strategy_type:
          type: string
          default: parent_child
          enum: [full_token, parent_child, semantic]
        metadata:
          type: object
        chunk_size:
          type: integer
          default: 500
          minimum: 100
        chunk_overlap:
          type: integer
          default: 50
          minimum: 0
        top_keywords:
          type: integer
          default: 10
          minimum: 1
          maximum: 50
    ChunkIndexAPIResponse:
      type: object
      properties:
        document_id:
          type: string
        user_id:
          type: string
        total_chunks:
          type: integer
        indexed_chunks:
          type: array
          items:
            type: object
            properties:
              chunk_id:
                type: string
              chunk_type:
                type: string
              keywords:
                type: array
                items:
                  type: string
              content:
                type: string
        request_id:
          type: string
    MorphIndexAPIRequest:
      type: object
      required: [document_id, content, user_id]
      properties:
        document_id:
          type: string
        content:
          type: string
        user_id:
          type: string
        strategy_type:
          type: string
          default: parent_child
          enum: [full_token, parent_child, semantic]
        chunk_size:
          type: integer
          default: 500
          minimum: 100
        chunk_overlap:
          type: integer
          default: 50
          minimum: 0
        source:
          type: string
          default: ""
        metadata:
          type: object
    MorphIndexAPIResponse:
      type: object
      properties:
        document_id:
          type: string
        user_id:
          type: string
        total_chunks:
          type: integer
        qdrant_indexed:
          type: integer
        es_indexed:
          type: integer
        indexed_chunks:
          type: array
          items:
            type: object
            properties:
              chunk_id:
                type: string
              chunk_type:
                type: string
              morph_keywords:
                type: array
                items:
                  type: string
              content:
                type: string
              char_start:
                type: integer
              char_end:
                type: integer
              chunk_index:
                type: integer
        request_id:
          type: string
...
    CreateAgentRequest:
      type: object
      required: [name, goal]
      properties:
        name:
          type: string
        goal:
          type: string
        tool_ids:
          type: array
          items:
            type: string
    CreateAgentResponse:
      type: object
      properties:
        agent_id:
          type: string
        name:
          type: string
        system_prompt:
          type: string
        tool_ids:
          type: array
          items:
            type: string
    UpdateAgentRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        system_prompt:
          type: string
          nullable: true
...
    CreateMiddlewareAgentRequest:
      type: object
      required: [name, goal]
      properties:
        name:
          type: string
        goal:
          type: string
        tool_ids:
          type: array
          items:
            type: string
        middlewares:
          type: array
          items:
            type: string
          description: "SummarizationMiddleware, PIIMiddleware, ToolRetryMiddleware, ModelCallLimitMiddleware, ModelFallbackMiddleware"
    UpdateMiddlewareAgentRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        system_prompt:
          type: string
          nullable: true
        middlewares:
          type: array
          items:
            type: string
...
    AutoBuildRequest:
      type: object
      required: [user_request, user_id]
      properties:
        user_request:
          type: string
          description: 자연어로 에이전트 요구사항 설명
        user_id:
          type: string
    AutoBuildReplyRequest:
      type: object
      required: [answer]
      properties:
        answer:
          type: string
    AutoBuildResponse:
      type: object
      properties:
        session_id:
          type: string
        status:
          type: string
          enum: [created, needs_clarification]
        agent_id:
          type: string
          nullable: true
        clarification_question:
          type: string
          nullable: true
    AutoBuildSessionStatusResponse:
      type: object
      properties:
        session_id:
          type: string
        status:
          type: string
        attempt_count:
          type: integer
        user_request:
          type: string
        created_agent_id:
          type: string
          nullable: true
...